+++ test/src/test-compat.c
@@ -62,7 +62,6 @@ static void openfile(const char *name) {
   else if(header) fprintf(fp, "%s", header);
 }
 
-
 static void printvector(double *v) {
   if(!v) fprintf(fp, "null ");
   fprintf(fp, "%12.6f %12.6f %12.6f ", v[0], v[1], v[2]);
@@ -72,8 +71,8 @@ static void printunitvector(double *v) {
   if(!v) fprintf(fp, "null ");
   else {
     double l = sqrt(v[0] * v[0] + v[1] * v[1] + v[2] * v[2]);
-    if(accuracy == 0) fprintf(fp, "%12.9f %12.9f %12.9f ", v[0] / l, v[1] / l, v[2] / l);
-    else fprintf(fp, "%9.6f %9.6f %9.6f ", v[0] / l, v[1] / l, v[2] / l);
+    if(accuracy == 0) fprintf(fp, "%14.11f %14.11f %14.11f ", v[0] / l, v[1] / l, v[2] / l);
+    else fprintf(fp, "%11.8f %11.8f %11.8f ", v[0] / l, v[1] / l, v[2] / l);
   }
 }
 
@@ -202,15 +201,15 @@ static void test_mean_star() {
 
   openfile("mean_star");
   if(is_ok(mean_star(2433282.42345905, 10.0, -40.0, 1, &ra, &dec)))
-    fprintf(fp, "1 %12.6f %12.6f ", ra, dec);
+    fprintf(fp, "1 %12.9f %12.8f ", ra, dec);
 
   openfile("mean_star");
   if(is_ok(mean_star(2433282.42345905, 19.0, 30.0, 1, &ra, &dec)))
-    fprintf(fp, "2 %12.6f %12.6f ", ra, dec);
+    fprintf(fp, "2 %12.9f %12.8f ", ra, dec);
 
   openfile("mean_star");
   if(is_ok(mean_star(2433282.42345905, 2.7, 68.3, 1, &ra, &dec)))
-    fprintf(fp, "3 %12.6f %12.6f ", ra, dec);
+    fprintf(fp, "3 %12.9f %12.8f ", ra, dec);
 }